chore(connect): remove bowser dep
