+++ packages/connect/package.json
@@ -56,7 +56,6 @@
         "@trezor/utxo-lib": "workspace:^1.0.4",
         "bignumber.js": "^9.1.1",
         "blakejs": "^1.2.1",
-        "bowser": "^2.11.0",
         "cross-fetch": "^3.1.5",
         "events": "^3.3.0",
         "parse-uri": "1.0.7",
