+++ packages/connect/src/data/udevInfo.ts
@@ -22,3 +22,18 @@ const info: UdevInfo = {
 };
 
 export const getUdevInfo = () => info;
+
+export const suggestUdevInstaller = (platform?: string) => {
+    const info = getUdevInfo();
+    // check if preferred field was already added
+    if (!info.packages.find(p => p.preferred)) {
+        if (platform) {
+            // override UdevInfo packages, add preferred field
+            info.packages = info.packages.map(p => ({
+                ...p,
+                preferred: p.platform.indexOf(platform) >= 0,
+            }));
+        }
+    }
+    return info;
+};
